backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
PyPDF2==3.0.1
Pillow==10.1.0
opencv-python==4.8.1.78
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
requests==2.31.0
aiohttp==3.9.1
boto3==1.34.0
botocore==1.34.0
slowapi==0.1.9
pytesseract==0.3.10
pdf2image==1.16.3
python-pptx==0.6.23
